- CHG: Reworked data importers/exporters to use the database for storing job parameters. - CHG: Importer/exporter jobs now set their own status when finished/failed. This makes them appear as done quicker. - CHG: Reduced frequency of background tasks that check whether import/export tasks are done. They still exist as a fallback to eventually set jobs to completed/failed if the job fails to do so itself.

+++ src/main/java/jhi/germinate/server/util/async/GenotypeExporter.java
@@ -2,19 +2,41 @@ package jhi.germinate.server.util.async;
 
 import jhi.flapjack.io.FlapjackFile;
 import jhi.flapjack.io.cmd.*;
-import jhi.germinate.server.util.StringUtils;
+import jhi.germinate.server.Database;
+import jhi.germinate.server.database.codegen.enums.DatasetExportJobsStatus;
+import jhi.germinate.server.database.codegen.tables.pojos.*;
+import jhi.germinate.server.database.codegen.tables.records.DatasetExportJobsRecord;
+import jhi.germinate.server.database.pojo.AdditionalExportFormat;
+import jhi.germinate.server.resource.ResourceUtils;
+import jhi.germinate.server.util.*;
 import jhi.germinate.server.util.hdf5.Hdf5ToFJTabbedConverter;
 import jhi.germinate.server.util.hdf5.*;
+import org.jooq.*;
+import org.jooq.impl.DSL;
 
+import java.io.File;
 import java.io.*;
 import java.net.URI;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystem;
+import java.nio.file.Files;
 import java.nio.file.*;
+import java.sql.*;
 import java.text.SimpleDateFormat;
 import java.time.*;
+import java.util.Date;
 import java.util.*;
 import java.util.concurrent.*;
-import java.util.stream.Collectors;
+import java.util.logging.Logger;
+
+import static jhi.germinate.server.database.codegen.tables.DatasetExportJobs.*;
+import static jhi.germinate.server.database.codegen.tables.Datasetmembers.*;
+import static jhi.germinate.server.database.codegen.tables.Datasets.*;
+import static jhi.germinate.server.database.codegen.tables.Germinatebase.*;
+import static jhi.germinate.server.database.codegen.tables.Groupmembers.*;
+import static jhi.germinate.server.database.codegen.tables.Mapdefinitions.*;
+import static jhi.germinate.server.database.codegen.tables.Markers.*;
+import static jhi.germinate.server.database.codegen.tables.Synonyms.*;
 
 /**
  * @author Sebastian Raubach
@@ -23,23 +45,22 @@ public class GenotypeExporter
 {
 	private static final SimpleDateFormat SDF = new SimpleDateFormat("yyyy-MM-dd-HH-mm-ss");
 
-	private File        folder;
-	private File        hdf5File;
-	private File        hdf5TransposedFile;
-	private File        mapFile;
-	private File        tabbedFile;
-	private File        germplasmFile;
-	private File        markersFile;
-	private File        headerFile;
-	private File        flapjackProjectFile;
-	private File        hapmapFile;
-	private File        identifierFile;
-	private File        zipFile;
-	private Set<String> germplasm;
-	private Set<String> markers;
-	private String      headers         = "";
-	private String      projectName;
-	private boolean     includeFlatText = true;
+	private File              folder;
+	private File              hdf5File;
+	private File              hdf5TransposedFile;
+	private File              mapFile;
+	private File              tabbedFile;
+	private File              flapjackProjectFile;
+	private File              hapmapFile;
+	private File              identifierFile;
+	private File              zipFile;
+	private Set<String>       germplasm;
+	private Set<String>       markers;
+	private String            headers         = "";
+	private String            projectName;
+	private boolean           includeFlatText = true;
+	private DatasetExportJobs exportJob;
+	private Datasets          dataset;
 
 	private final CountDownLatch latch;
 
@@ -57,106 +78,234 @@ public class GenotypeExporter
 	public static void main(String[] args)
 		throws IOException
 	{
-		int i = 0;
 		GenotypeExporter exporter = new GenotypeExporter();
-		exporter.hdf5File = new File(args[i++]);
+		Database.init(args[0], args[1], args[2], args[3], args[4], false);
+		Integer jobId = Integer.parseInt(args[5]);
 
-		File potential = new File(exporter.hdf5File.getParentFile(), "transposed-" + exporter.hdf5File.getName());
+		try (Connection conn = Database.getConnection())
+		{
+			DSLContext context = Database.getContext(conn);
 
-		if (potential.exists() && potential.isFile())
-			exporter.hdf5TransposedFile = potential;
+			exporter.exportJob = context.selectFrom(DATASET_EXPORT_JOBS).where(DATASET_EXPORT_JOBS.ID.eq(jobId)).fetchAnyInto(DatasetExportJobs.class);
+			exporter.dataset = context.selectFrom(DATASETS).where(DATASETS.ID.eq(exporter.exportJob.getDatasetIds()[0])).fetchAnyInto(Datasets.class);
 
-		exporter.folder = new File(args[i++]);
-		exporter.projectName = args[i++];
-		String additionalFormats = args[i++];
-		String[] parts = additionalFormats.split(",");
-		List<AdditionalExportFormat> formats = Arrays.stream(parts)
-													 .map(p -> {
-														 try
-														 {
-															 return AdditionalExportFormat.valueOf(p);
-														 }
-														 catch (Exception e)
-														 {
-															 return null;
-														 }
-													 })
-													 .filter(Objects::nonNull)
-													 .collect(Collectors.toList());
+			exporter.hdf5File = new File(new File(new File(exporter.exportJob.getJobConfig().getBaseFolder(), "data"), "genotypes"), exporter.dataset.getSourceFile());
 
-		exporter.tabbedFile = new File(exporter.folder, exporter.projectName + ".txt");
-		exporter.zipFile = new File(exporter.folder, exporter.projectName + "-" + SDF.format(new Date()) + ".zip");
+			File potential = new File(exporter.hdf5File.getParentFile(), "transposed-" + exporter.hdf5File.getName());
 
-		File germplasmFile = new File(exporter.folder, exporter.projectName + ".germplasm");
-		File markersFile = new File(exporter.folder, exporter.projectName + ".markers");
-		File headersFile = new File(exporter.folder, exporter.projectName + ".header");
-		File mapFile = new File(exporter.folder, exporter.projectName + ".map");
-		File identifierFile = new File(exporter.folder, exporter.projectName + ".identifiers");
+			if (potential.exists() && potential.isFile()) exporter.hdf5TransposedFile = potential;
 
-		if (germplasmFile.exists() && germplasmFile.isFile())
-			exporter.germplasmFile = germplasmFile;
-		if (markersFile.exists() && markersFile.isFile())
-			exporter.markersFile = markersFile;
-		if (headersFile.exists() && headersFile.isFile())
-			exporter.headerFile = headersFile;
-		if (identifierFile.exists() && identifierFile.isFile())
-			exporter.identifierFile = identifierFile;
-		if (mapFile.exists() && mapFile.isFile())
-			exporter.mapFile = mapFile;
-		else
-			exporter.mapFile = null;
-		if (formats.contains(AdditionalExportFormat.flapjack))
-			exporter.flapjackProjectFile = new File(exporter.folder, exporter.projectName + ".flapjack");
-		if (formats.contains(AdditionalExportFormat.hapmap))
-			exporter.hapmapFile = new File(exporter.folder, exporter.projectName + ".hapmap");
+			exporter.folder = new File(new File(exporter.exportJob.getJobConfig().getBaseFolder(), "async"), exporter.exportJob.getUuid());
+			exporter.projectName = "dataset-" + exporter.dataset.getId();
 
-		exporter.includeFlatText = formats.contains(AdditionalExportFormat.text);
+			if (exporter.exportJob.getJobConfig().getSubsetId() != null)
+				exporter.projectName += "-map-" + exporter.exportJob.getJobConfig().getSubsetId();
 
-		exporter.run();
+			List<AdditionalExportFormat> formats = Arrays.asList(exporter.exportJob.getJobConfig().getFileTypes());
+
+			exporter.tabbedFile = new File(exporter.folder, exporter.projectName + ".txt");
+			exporter.zipFile = new File(exporter.folder, exporter.projectName + "-" + SDF.format(new Date()) + ".zip");
+
+			exporter.mapFile = new File(exporter.folder, exporter.projectName + ".map");
+			exporter.identifierFile = new File(exporter.folder, exporter.projectName + ".identifiers");
+
+			if (formats.contains(AdditionalExportFormat.flapjack))
+				exporter.flapjackProjectFile = new File(exporter.folder, exporter.projectName + ".flapjack");
+			if (formats.contains(AdditionalExportFormat.hapmap))
+				exporter.hapmapFile = new File(exporter.folder, exporter.projectName + ".hapmap");
+
+			exporter.includeFlatText = formats.contains(AdditionalExportFormat.text);
+
+			exporter.run();
+
+			DatasetExportJobsRecord record = context.selectFrom(DATASET_EXPORT_JOBS).where(DATASET_EXPORT_JOBS.ID.eq(jobId)).fetchAny();
+			record.setStatus(DatasetExportJobsStatus.completed);
+			record.setResultSize(exporter.zipFile.length());
+			record.store(DATASET_EXPORT_JOBS.STATUS, DATASET_EXPORT_JOBS.RESULT_SIZE);
+		}
+		catch (Exception e)
+		{
+			e.printStackTrace();
+			try (Connection conn = Database.getConnection())
+			{
+				DSLContext context = Database.getContext(conn);
+
+				DatasetExportJobsRecord record = context.selectFrom(DATASET_EXPORT_JOBS).where(DATASET_EXPORT_JOBS.ID.eq(jobId)).fetchAny();
+				record.setStatus(DatasetExportJobsStatus.failed);
+				record.store(DATASET_EXPORT_JOBS.STATUS);
+			}
+			catch (Exception ee)
+			{
+				ee.printStackTrace();
+				Logger.getLogger("").severe(ee.getMessage());
+			}
+		}
 	}
 
 	private void init()
-		throws IOException
+		throws IOException, SQLException
 	{
-		if (germplasmFile != null)
+		try (Connection conn = Database.getConnection())
 		{
-			germplasm = new LinkedHashSet<>(Files.readAllLines(germplasmFile.toPath()));
-			germplasmFile.delete();
+			DSLContext context = Database.getContext(conn);
+
+			// Write the map to a file
+			if (exportJob.getJobConfig().getSubsetId() != null)
+			{
+				mapFile = new File(folder, projectName + ".map");
+				SelectConditionStep<Record3<String, String, Double>> query = context.select(MARKERS.MARKER_NAME, MAPDEFINITIONS.CHROMOSOME, MAPDEFINITIONS.DEFINITION_START)
+																					.from(MAPDEFINITIONS)
+																					.leftJoin(MARKERS).on(MARKERS.ID.eq(MAPDEFINITIONS.MARKER_ID))
+																					.leftJoin(DATASETMEMBERS).on(DATASETMEMBERS.FOREIGN_ID.eq(MARKERS.ID).and(DATASETMEMBERS.DATASETMEMBERTYPE_ID.eq(1)))
+																					.where(DATASETMEMBERS.DATASET_ID.eq(exportJob.getDatasetIds()[0]))
+																					.and(MAPDEFINITIONS.MAP_ID.eq(exportJob.getJobConfig().getSubsetId()));
+
+				try (PrintWriter bw = new PrintWriter(new BufferedWriter(new OutputStreamWriter(new FileOutputStream(mapFile), StandardCharsets.UTF_8)));
+					 Cursor<? extends Record> cursor = query.fetchLazy())
+				{
+					bw.write("# fjFile = MAP" + ResourceUtils.CRLF);
+
+					ResourceUtils.exportToFileStreamed(bw, cursor, false, null);
+				}
+			}
+
+			// Get the germplasm
+			if (exportJob.getJobConfig().getyGroupIds() != null || exportJob.getJobConfig().getyIds() != null)
+			{
+				Set<String> result = new LinkedHashSet<>();
+
+				if (!CollectionUtils.isEmpty(exportJob.getJobConfig().getyIds()))
+				{
+					result.addAll(context.selectDistinct(GERMINATEBASE.NAME)
+										 .from(GERMINATEBASE)
+										 .where(GERMINATEBASE.ID.in(exportJob.getJobConfig().getyIds()))
+										 .fetchInto(String.class));
+				}
+				if (!CollectionUtils.isEmpty(exportJob.getJobConfig().getyGroupIds()))
+				{
+					result.addAll(context.selectDistinct(GERMINATEBASE.NAME)
+										 .from(GERMINATEBASE)
+										 .leftJoin(GROUPMEMBERS).on(GROUPMEMBERS.FOREIGN_ID.eq(GERMINATEBASE.ID))
+										 .where(GROUPMEMBERS.GROUP_ID.in(exportJob.getJobConfig().getyGroupIds()))
+										 .fetchInto(String.class));
+				}
+
+				this.germplasm = result;
+			}
+
+			// Get the markers
+			if (exportJob.getJobConfig().getxGroupIds() != null || exportJob.getJobConfig().getxIds() != null)
+			{
+				Set<String> result = new LinkedHashSet<>();
+				if (!CollectionUtils.isEmpty(exportJob.getJobConfig().getxIds()))
+				{
+					result.addAll(context.selectDistinct(MARKERS.MARKER_NAME)
+										 .from(MARKERS)
+										 .where(MARKERS.ID.in(exportJob.getJobConfig().getxIds()))
+										 .fetchInto(String.class));
+				}
+
+				if (!CollectionUtils.isEmpty(exportJob.getJobConfig().getxGroupIds()))
+				{
+					result.addAll(context.selectDistinct(MARKERS.MARKER_NAME)
+										 .from(MARKERS)
+										 .leftJoin(GROUPMEMBERS).on(GROUPMEMBERS.FOREIGN_ID.eq(MARKERS.ID))
+										 .where(GROUPMEMBERS.GROUP_ID.in(exportJob.getJobConfig().getxGroupIds()))
+										 .fetchInto(String.class));
+				}
+
+				if (exportJob.getJobConfig().getSubsetId() != null)
+				{
+					// Only keep those that are actually on the map
+					result.retainAll(context.selectDistinct(MARKERS.MARKER_NAME)
+											.from(MARKERS)
+											.leftJoin(MAPDEFINITIONS).on(MAPDEFINITIONS.MARKER_ID.eq(MARKERS.ID))
+											.where(MAPDEFINITIONS.MAP_ID.eq(exportJob.getJobConfig().getSubsetId()))
+											.fetchInto(String.class));
+				}
+
+				markers = result;
+			}
+
+			// Write the identifiers file
+			try (BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(new FileOutputStream(identifierFile), StandardCharsets.UTF_8)))
+			{
+				bw.write("# fjFile = PHENOTYPE");
+				bw.newLine();
+				bw.write("\tPUID\tSource Material Name\tSource Material PUID\tSynonyms");
+
+				jhi.germinate.server.database.codegen.tables.Germinatebase g = GERMINATEBASE.as("g");
+				Field<String> childName = GERMINATEBASE.NAME.as("childName");
+				Field<String> childPuid = GERMINATEBASE.PUID.as("childPuid");
+				Field<String> parentName = g.NAME.as("parentName");
+				Field<String> parentPuid = g.PUID.as("parentPuid");
+				Field<String[]> synonyms = SYNONYMS.SYNONYMS_.as("synonyms");
+
+				SelectJoinStep<Record5<String, String, String, String, String[]>> step = context.select(
+					childName,
+					childPuid,
+					parentName,
+					parentPuid,
+					synonyms
+				).from(GERMINATEBASE.leftJoin(g).on(g.ID.eq(GERMINATEBASE.ENTITYPARENT_ID))
+									.leftJoin(SYNONYMS).on(SYNONYMS.SYNONYMTYPE_ID.eq(1).and(SYNONYMS.FOREIGN_ID.eq(GERMINATEBASE.ID))));
+
+				// Restrict to the requested germplasm (if any)
+				if (!CollectionUtils.isEmpty(germplasm))
+					step.where(GERMINATEBASE.NAME.in(germplasm));
+					// Otherwise, restrict it to everything in this dataset
+				else
+					step.where(DSL.exists(DSL.selectOne()
+											 .from(DATASETMEMBERS)
+											 .where(DATASETMEMBERS.FOREIGN_ID.eq(GERMINATEBASE.ID)
+																			 .and(DATASETMEMBERS.DATASETMEMBERTYPE_ID.eq(2))
+																			 .and(DATASETMEMBERS.DATASET_ID.eq(exportJob.getDatasetIds()[0])))));
+
+				// Get only the ones where there's either an entity parent or the PUID or the synonyms are't null, otherwise we're wasting space in the file
+				step.where(GERMINATEBASE.ENTITYPARENT_ID.isNotNull().or(GERMINATEBASE.PUID.isNotNull()).or(SYNONYMS.SYNONYMS_.isNotNull()));
+
+				step.forEach(r -> {
+					try
+					{
+						bw.newLine();
+						bw.write(r.get(childName) + "\t");
+						bw.write((r.get(childPuid) == null ? "" : r.get(childPuid)) + "\t");
+						bw.write((r.get(parentName) == null ? "" : r.get(parentName)) + "\t");
+						bw.write((r.get(parentPuid) == null ? "" : r.get(parentPuid)) + "\t");
+						bw.write((r.get(synonyms)) == null ? "" : r.get(synonyms).toString());
+					}
+					catch (IOException e)
+					{
+						e.printStackTrace();
+					}
+				});
+			}
+			catch (IOException e)
+			{
+				e.printStackTrace();
+			}
 		}
 
-		if (markersFile != null)
-		{
-			markers = new LinkedHashSet<>(Files.readAllLines(markersFile.toPath()));
-			markersFile.delete();
-		}
-
-		if (headerFile != null)
-		{
-			headers = String.join("\n", Files.readAllLines(headerFile.toPath())) + "\n";
-			headerFile.delete();
-		}
+		headers = exportJob.getJobConfig().getFileHeaders();
 	}
 
 	private void zip(FileSystem fs, File f, boolean delete)
 		throws IOException
 	{
 		Files.copy(f.toPath(), fs.getPath("/" + f.getName()), StandardCopyOption.REPLACE_EXISTING);
-		if (delete)
-			f.delete();
+		if (delete) f.delete();
 	}
 
 	private List<String> run()
-		throws IOException
+		throws IOException, SQLException
 	{
 		init();
 
 		// Make sure it doesn't exist
-		if (zipFile.exists())
-			zipFile.delete();
+		if (zipFile.exists()) zipFile.delete();
 
 		String prefix = zipFile.getAbsolutePath().replace("\\", "/");
-		if (prefix.startsWith("/"))
-			prefix = prefix.substring(1);
+		if (prefix.startsWith("/")) prefix = prefix.substring(1);
 		URI uri = URI.create("jar:file:/" + prefix);
 		Map<String, Object> env = new HashMap<>();
 		// We need to write this to a temp file, because genotypic data is potentially much larger than the memory and by default ZipFileSystem
@@ -234,10 +383,8 @@ public class GenotypeExporter
 			latch.await();
 
 			// Now delete the remaining files
-			if (mapFile != null)
-				mapFile.delete();
-			if (tabbedFile != null)
-				tabbedFile.delete();
+			if (mapFile != null) mapFile.delete();
+			if (tabbedFile != null) tabbedFile.delete();
 
 			Duration duration = Duration.between(start, Instant.now());
 			System.out.println("DURATION: " + duration);
@@ -308,8 +455,7 @@ public class GenotypeExporter
 
 				Files.move(tempTarget.toPath(), flapjackProjectFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
 
-				if (includeFlatText)
-					zip(fs, tabbedFile, true);
+				if (includeFlatText) zip(fs, tabbedFile, true);
 
 				zip(fs, flapjackProjectFile, true);
 			}
@@ -333,14 +479,10 @@ public class GenotypeExporter
 				Map<String, Hdf5ToHapmapConverter.MarkerPosition> map = new HashMap<>();
 				if (mapFile != null)
 				{
-					Files.readAllLines(mapFile.toPath())
-						 .stream()
-						 .skip(1)
-						 .filter(l -> !StringUtils.isEmpty(l))
-						 .forEachOrdered(m -> {
-							 String[] parts = m.split("\t");
-							 map.put(parts[0], new Hdf5ToHapmapConverter.MarkerPosition(parts[1], Integer.toString((int) Math.round(Double.parseDouble(parts[2])))));
-						 });
+					Files.readAllLines(mapFile.toPath()).stream().skip(1).filter(l -> !StringUtils.isEmpty(l)).forEachOrdered(m -> {
+						String[] parts = m.split("\t");
+						map.put(parts[0], new Hdf5ToHapmapConverter.MarkerPosition(parts[1], Integer.toString((int) Math.round(Double.parseDouble(parts[2])))));
+					});
 				}
 
 				Path hapmapPath = fs.getPath("/" + hapmapFile.getName());
@@ -349,8 +491,7 @@ public class GenotypeExporter
 
 				if (hdf5TransposedFile != null)
 					converter = new Hdf5TransposedToHapmapConverter(hdf5TransposedFile.toPath(), germplasm, markers, map, hapmapPath);
-				else
-					converter = new Hdf5ToHapmapConverter(hdf5File.toPath(), germplasm, markers, map, hapmapPath);
+				else converter = new Hdf5ToHapmapConverter(hdf5File.toPath(), germplasm, markers, map, hapmapPath);
 
 				converter.extractData(null);
 			}
